# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cidades e Clientes API",
    "description": "\n    Desafio de desenvolvimento de uma Web API Rest da Compasso UOL.\n\n    Desenvolvido por: Alexsander Apulinario Martins Alves.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/cidade": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          },
          {
            "name": "estado",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/cidade/cadastrar": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/cliente/cadastrar": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nomeCompleto": {
                  "example": "any"
                },
                "sexo": {
                  "example": "any"
                },
                "dataNascimento": {
                  "example": "any"
                },
                "cidade": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/cliente": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "novo_nome_completo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "novo_nome_completo",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cidade": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                },
                "nomeCompleto": {
                  "example": "any"
                },
                "sexo": {
                  "example": "any"
                },
                "dataNascimento": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {}
}